+++ src/nvim/insexpand.c
@@ -43,6 +43,7 @@
 #include "nvim/state.h"
 #include "nvim/strings.h"
 #include "nvim/tag.h"
+#include "nvim/textformat.h"
 #include "nvim/ui.h"
 #include "nvim/undo.h"
 #include "nvim/vim.h"
