+++ src/nvim/textformat.h
@@ -1,0 +1,12 @@
+#ifndef NVIM_TEXTFORMAT_H
+#define NVIM_TEXTFORMAT_H
+
+#include <stdbool.h>
+
+#include "nvim/normal.h"  // for oparg_T
+#include "nvim/pos.h"  // for linenr_T
+
+#ifdef INCLUDE_GENERATED_DECLARATIONS
+# include "textformat.h.generated.h"
+#endif
+#endif  // NVIM_TEXTFORMAT_H
